+++ mongod/create.go
@@ -16,7 +16,7 @@ func (c *appRepository) CreateDocument(collectionName string, document bson.D, u
 
 	// Set the updated_at field to the current time
 	document = append(document, bson.E{Key: "updated_at", Value: time.Now()})
-	fmt.Println("DOCUMENT: ", document[1])
+	logrus.Debugln("Document: ", document[1])
 	uploadResult, err := collection.InsertOne(ctx, document)
 	if err != nil {
 		if mongoErr, ok := err.(mongo.WriteException); ok {
@@ -36,17 +36,17 @@ func (c *appRepository) CreateDocument(collectionName string, document bson.D, u
 // CreateChannel creates a new channel document
 func (c *appRepository) CreateChannel(channelName string, ctx context.Context) (interface{}, error) {
 	document := bson.D{{Key: "channel_name", Value: channelName}}
-	return c.CreateDocument("apps", document, "channel_name_sort_by_asc_created", "channel", ctx)
+	return c.CreateDocument("apps_meta", document, "channel_name_sort_by_asc_created", "channel", ctx)
 }
 
 // CreatePlatform creates a new platform document
 func (c *appRepository) CreatePlatform(platformName string, ctx context.Context) (interface{}, error) {
 	document := bson.D{{Key: "platform_name", Value: platformName}}
-	return c.CreateDocument("apps", document, "platform_name_sort_by_asc_created", "platform", ctx)
+	return c.CreateDocument("apps_meta", document, "platform_name_sort_by_asc_created", "platform", ctx)
 }
 
 // CreateArch creates a new arch document
 func (c *appRepository) CreateArch(archID string, ctx context.Context) (interface{}, error) {
 	document := bson.D{{Key: "arch_id", Value: archID}}
-	return c.CreateDocument("apps", document, "arch_id_sort_by_asc_created", "arch", ctx)
+	return c.CreateDocument("apps_meta", document, "arch_id_sort_by_asc_created", "arch", ctx)
 }